requirements.txt
aiofiles==0.6.0
azure-core==1.13.0
azure-storage-blob==12.8.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
databases==0.4.3
ecdsa==0.14.1
fastapi==0.63.0
h11==0.12.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.0
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.1
pymongo==3.11.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-jose==3.2.0
python-multipart==0.0.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
SQLAlchemy==1.3.24
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
